feat(lambda): changes to controllers, services, apiErrorHandler, config/passport, event.json
Changes to controllers, services, apiErrorHandler, config/passport, event.json. Controllers are now
taking the userId from the req.user which is automatically populated from the config/passport.js
file that takes the user object from the Authorization Bearer JWT Token. Services are refactored as
well not to expect a token and the jwt.verify method is removed as it is no longer used. Changes to
apiErrorHandler error middleware which is now having a fallback statusCode of 500 in case the error
does not have a status code. Event.json is now modified to fire a delete event but not really a
relevant change as there is no DynamoDB configuration locally anyway.

BREAKING CHANGE: No longer token in headers. Now instead we are sending the JWT in the Authorization
header.